--- pkg/sypgp/fetch_test.go
@@ -1,90 +1,0 @@
-// Copyright (c) 2018, Sylabs Inc. All rights reserved.
-// This software is licensed under a 3-clause BSD license. Please consult the
-// LICENSE.md file distributed with the sources of this project regarding your
-// rights to use or distribute this software.
-
-package sypgp
-
-import (
-	"log"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/sylabs/singularity/internal/pkg/test"
-	"golang.org/x/crypto/openpgp"
-	"golang.org/x/crypto/openpgp/armor"
-)
-
-type mockServer struct {
-	code int
-	el   openpgp.EntityList
-}
-
-func (ms *mockServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(ms.code)
-	if ms.code == http.StatusOK {
-		w.Header().Set("Content-Type", "application/pgp-keys")
-
-		wr, err := armor.Encode(w, openpgp.PublicKeyType, nil)
-		if err != nil {
-			log.Fatalf("failed to get encoder: %v", err)
-		}
-		defer wr.Close()
-
-		for _, e := range ms.el {
-			if err = e.Serialize(wr); err != nil {
-				log.Fatalf("failed to serialize entity: %v", err)
-			}
-		}
-	}
-}
-
-func TestFetchPubkey(t *testing.T) {
-	ms := &mockServer{}
-	srv := httptest.NewServer(ms)
-	defer srv.Close()
-
-	fp := string(testEntity.PrimaryKey.Fingerprint[:])
-
-	tests := []struct {
-		name        string
-		code        int
-		el          openpgp.EntityList
-		fingerprint string
-		uri         string
-		authToken   string
-		wantErr     bool
-	}{
-		{"Success", http.StatusOK, openpgp.EntityList{testEntity}, fp, srv.URL, "", false},
-		{"SuccessToken", http.StatusOK, openpgp.EntityList{testEntity}, fp, srv.URL, "token", false},
-		{"NoKeys", http.StatusOK, openpgp.EntityList{}, fp, srv.URL, "token", true},
-		{"TwoKeys", http.StatusOK, openpgp.EntityList{testEntity, testEntity}, fp, srv.URL, "token", true},
-		{"BadURL", http.StatusOK, openpgp.EntityList{testEntity}, fp, ":", "", true},
-		{"NotFound", http.StatusNotFound, nil, fp, srv.URL, "", true},
-		{"Unauthorized", http.StatusUnauthorized, nil, fp, srv.URL, "", true},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, test.WithoutPrivilege(func(t *testing.T) {
-			ms.code = tt.code
-			ms.el = tt.el
-
-			el, err := FetchPubkey(tt.fingerprint, tt.uri, tt.authToken, false)
-			if (err != nil) != tt.wantErr {
-				t.Fatalf("unexpected error: %v", err)
-				return
-			}
-
-			if !tt.wantErr {
-				if len(el) != 1 {
-					t.Fatalf("unexpected number of entities returned: %v", len(el))
-				}
-				for i := range tt.el {
-					if fp := el[i].PrimaryKey.Fingerprint; fp != tt.el[i].PrimaryKey.Fingerprint {
-						t.Errorf("fingerprint mismatch: %v / %v", fp, tt.el[i].PrimaryKey.Fingerprint)
-					}
-				}
-			}
-		}))
-	}
-}
