+++ pkg/sypgp/sypgp_test.go
@@ -6,12 +6,17 @@
 package sypgp
 
 import (
+	"encoding/hex"
 	"log"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"testing"
 
+	"github.com/sylabs/singularity/internal/pkg/test"
 	"github.com/sylabs/singularity/pkg/util/user-agent"
 	"golang.org/x/crypto/openpgp"
+	"golang.org/x/crypto/openpgp/armor"
 )
 
 const (
@@ -23,6 +28,168 @@ const (
 var (
 	testEntity *openpgp.Entity
 )
+
+type mockPKSLookup struct {
+	code int
+	el   openpgp.EntityList
+}
+
+func (ms *mockPKSLookup) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	w.WriteHeader(ms.code)
+	if ms.code == http.StatusOK {
+		w.Header().Set("Content-Type", "application/pgp-keys")
+
+		wr, err := armor.Encode(w, openpgp.PublicKeyType, nil)
+		if err != nil {
+			log.Fatalf("failed to get encoder: %v", err)
+		}
+		defer wr.Close()
+
+		for _, e := range ms.el {
+			if err = e.Serialize(wr); err != nil {
+				log.Fatalf("failed to serialize entity: %v", err)
+			}
+		}
+	}
+}
+
+func TestSearchPubkey(t *testing.T) {
+	ms := &mockPKSLookup{}
+	srv := httptest.NewServer(ms)
+	defer srv.Close()
+
+	tests := []struct {
+		name      string
+		code      int
+		el        openpgp.EntityList
+		search    string
+		uri       string
+		authToken string
+		wantErr   bool
+	}{
+		{"Success", http.StatusOK, openpgp.EntityList{testEntity}, "search", srv.URL, "", false},
+		{"SuccessToken", http.StatusOK, openpgp.EntityList{testEntity}, "search", srv.URL, "token", false},
+		{"Unauthorized", http.StatusUnauthorized, nil, "search", srv.URL, "", true},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, test.WithoutPrivilege(func(t *testing.T) {
+			ms.code = tt.code
+			ms.el = tt.el
+
+			if err := SearchPubkey(tt.search, tt.uri, tt.authToken); (err != nil) != tt.wantErr {
+				t.Fatalf("got err %v, want error %v", err, tt.wantErr)
+			}
+		}))
+	}
+}
+
+func TestFetchPubkey(t *testing.T) {
+	ms := &mockPKSLookup{}
+	srv := httptest.NewServer(ms)
+	defer srv.Close()
+
+	fp := hex.EncodeToString(testEntity.PrimaryKey.Fingerprint[:])
+
+	tests := []struct {
+		name        string
+		code        int
+		el          openpgp.EntityList
+		fingerprint string
+		uri         string
+		authToken   string
+		wantErr     bool
+	}{
+		{"Success", http.StatusOK, openpgp.EntityList{testEntity}, fp, srv.URL, "", false},
+		{"SuccessToken", http.StatusOK, openpgp.EntityList{testEntity}, fp, srv.URL, "token", false},
+		{"NoKeys", http.StatusOK, openpgp.EntityList{}, fp, srv.URL, "token", true},
+		{"TwoKeys", http.StatusOK, openpgp.EntityList{testEntity, testEntity}, fp, srv.URL, "token", true},
+		{"BadURL", http.StatusOK, openpgp.EntityList{testEntity}, fp, ":", "", true},
+		{"NotFound", http.StatusNotFound, nil, fp, srv.URL, "", true},
+		{"Unauthorized", http.StatusUnauthorized, nil, fp, srv.URL, "", true},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, test.WithoutPrivilege(func(t *testing.T) {
+			ms.code = tt.code
+			ms.el = tt.el
+
+			el, err := FetchPubkey(tt.fingerprint, tt.uri, tt.authToken, false)
+			if (err != nil) != tt.wantErr {
+				t.Fatalf("unexpected error: %v", err)
+				return
+			}
+
+			if !tt.wantErr {
+				if len(el) != 1 {
+					t.Fatalf("unexpected number of entities returned: %v", len(el))
+				}
+				for i := range tt.el {
+					if fp := el[i].PrimaryKey.Fingerprint; fp != tt.el[i].PrimaryKey.Fingerprint {
+						t.Errorf("fingerprint mismatch: %v / %v", fp, tt.el[i].PrimaryKey.Fingerprint)
+					}
+				}
+			}
+		}))
+	}
+}
+
+type mockPKSAdd struct {
+	t       *testing.T
+	keyText string
+	code    int
+}
+
+func (m *mockPKSAdd) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if m.code != http.StatusOK {
+		w.WriteHeader(m.code)
+		return
+	}
+
+	if got, want := r.Header.Get("Content-Type"), "application/x-www-form-urlencoded"; got != want {
+		m.t.Errorf("got content type %v, want %v", got, want)
+	}
+
+	if err := r.ParseForm(); err != nil {
+		m.t.Fatalf("failed to parse form: %v", err)
+	}
+	if got, want := r.Form.Get("keytext"), m.keyText; got != want {
+		m.t.Errorf("got key text %v, want %v", got, want)
+	}
+}
+
+func TestPushPubkey(t *testing.T) {
+	keyText, err := serializeEntity(testEntity, openpgp.PublicKeyType)
+	if err != nil {
+		t.Fatalf("failed to serialize entity: %v", err)
+	}
+
+	ms := &mockPKSAdd{
+		t:       t,
+		keyText: keyText,
+	}
+	srv := httptest.NewServer(ms)
+	defer srv.Close()
+
+	tests := []struct {
+		name      string
+		authToken string
+		code      int
+		wantErr   bool
+	}{
+		{"Success", "", http.StatusOK, false},
+		{"SuccessToken", "token", http.StatusOK, false},
+		{"Unauthorized", "", http.StatusUnauthorized, true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			ms.code = tt.code
+
+			if err := PushPubkey(testEntity, srv.URL, tt.authToken); (err != nil) != tt.wantErr {
+				t.Fatalf("got err %v, want error %v", err, tt.wantErr)
+			}
+		})
+	}
+}
 
 func TestMain(m *testing.M) {
 	useragent.InitValue("singularity", "3.0.0-alpha.1-303-gaed8d30-dirty")
